// connectedhomeip/connectedhomeip/third_party/zap/repo/src-shared/types/uc-component-types.ts
/**
 *
 *    Copyright (c) 2020 Silicon Labs
 *
 *    Licensed under the Apache License, Version 2.0 (the "License");
 *    you may not use this file except in compliance with the License.
 *    You may obtain a copy of the License at
 *
 *        http://www.apache.org/licenses/LICENSE-2.0
 *
 *    Unless required by applicable law or agreed to in writing, software
 *    distributed under the License is distributed on an "AS IS" BASIS,
 *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *    See the License for the specific language governing permissions and
 *    limitations under the License.
 */

import * as http from 'http-status-codes'

export interface UcComponentIds {
  componentIds: string[]
  clusterId: number
  clusterLabel?: string
  side: string // 'client' or 'server'
}

export interface UcComponentUpdateResponseWrapper {
  status: http.StatusCodes
  id: string
  data: string
}

export interface UcComponentUpdateResponse {
  componentId?: string
}