// connectedhomeip/connectedhomeip/third_party/zap/repo/src-electron/types/db-mapping-types.ts
/**
 *
 *    Copyright (c) 2020 Silicon Labs
 *
 *    Licensed under the Apache License, Version 2.0 (the "License");
 *    you may not use this file except in compliance with the License.
 *    You may obtain a copy of the License at
 *
 *        http://www.apache.org/licenses/LICENSE-2.0
 *
 *    Unless required by applicable law or agreed to in writing, software
 *    distributed under the License is distributed on an "AS IS" BASIS,
 *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *    See the License for the specific language governing permissions and
 *    limitations under the License.
 */

// This file is generated via ./src-script/regen-ts-type.js.  Do not hand edit this file!

export interface SessionType {
  sessionId: number
  sessionKey: string
  creationTime: number
  dirty: boolean
}